refactor(Protocol_*): replace C-style casts and modernize protocol implementations
- Protocol_HTTP: Replace C-style casts with static_cast/reinterpret_cast across
  api_return, common_date, httpv1_client/server, headersparser, response,
  streamdecoder/encoder_url, websocket_framecontent, chunked_subparser
- Protocol_HTTP: Initialize tm structs with brace initialization in common_date
- Protocol_HTTP: Use brace initialization for addrinfo and sockaddr_in structs
- Protocol_MIME: Replace C-style casts with static_cast/reinterpret_cast in
  mime_message and mime_sub_header, initialize structs with brace init
- Protocol_FastRPC1/FastRPC3: Replace C-style casts with static_cast
- Protocol_LineRecv: Replace C-style cast with static_cast

+++ Mantids30/Protocol_HTTP/httpv1_server_headersparser.cpp
@@ -43,10 +43,10 @@ void HTTP::HTTPv1_Server::parseHostHeader()
 void HTTP::HTTPv1_Server::parsePort(const string &portStr)
 {
     char *endptr;
-    unsigned long port = (uint16_t) strtoul(portStr.c_str(), &endptr, 10);
+    uint16_t port = static_cast<uint16_t>(strtoul(portStr.c_str(), &endptr, 10));
     if (*endptr == '\0' && port <= 65535) // Check if conversion was successful and port is valid
     {
-        clientRequest.virtualPort = (uint16_t) port;
+        clientRequest.virtualPort = port;
     }
     else
     {
